Add start of rooms
This creates rooms. This is really far from doing what it should, but
it's the basics. This lowers our test coverage an incredible amount, as
most classes have been moved around.
